+++ lib/src/cross_axis_scroll.dart
@@ -1,6 +1,4 @@
 import 'dart:async';
-import 'dart:io';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'cross_scroll_bar.dart';
@@ -31,24 +29,11 @@ class CrossScroll extends StatefulWidget {
       this.horizontalScroll,
       this.horizontalScrollController,
       this.verticalScrollController,
-      this.idleColor,
+      this.dimColor,
+      this.normalColor,
       this.hoverColor,
-      this.horizontalBar = const CrossScrollBar(
-          isAlwaysShown: true,
-          trackVisibility: false,
-          radius: Radius.elliptical(8, 8),
-          interactive: false,
-          hoverThickness: 16,
-          thickness: 8,
-          showTrackOnHover: false),
-      this.verticalBar = const CrossScrollBar(
-          isAlwaysShown: true,
-          trackVisibility: false,
-          radius: Radius.elliptical(8, 8),
-          interactive: false,
-          hoverThickness: 16,
-          thickness: 8,
-          showTrackOnHover: false),
+      this.horizontalBar=const CrossScrollBar() ,
+      this.verticalBar =const CrossScrollBar() ,
       Key? key})
       : super(key: key);
   final Widget? child;
@@ -62,8 +47,6 @@ class CrossScroll extends StatefulWidget {
   /// An object that can be used to control the position to which this scroll
   /// view is scrolled.
   ///
-  /// Must be null if [primary] is true.
-  ///
   /// A [ScrollController] serves several purposes. It can be used to control
   /// the initial scroll position (see [ScrollController.initialScrollOffset]).
   /// It can be used to control whether the scroll view should automatically
@@ -76,8 +59,6 @@ class CrossScroll extends StatefulWidget {
 
   /// An object that can be used to control the position to which this scroll
   /// view is scrolled.
-  ///
-  /// Must be null if [primary] is true.
   ///
   /// A [ScrollController] serves several purposes. It can be used to control
   /// the initial scroll position (see [ScrollController.initialScrollOffset]).
@@ -96,11 +77,15 @@ class CrossScroll extends StatefulWidget {
   final CrossScrollBar? horizontalBar;
 
   ///Shown this color when you hovered over a thumb
+  ///
+  ///when [hoverColor] is null and [CrossScrollBar(thumb:ScrollThumb.hoverShow)] then [normalColor] will be assign to [hoverColor] if [normalColor] not null.
   final Color? hoverColor;
 
   ///Normal thumb color
-  ///By default it will get [Theme.of(context).highlightColor.withOpacity(0.9)]
-  final Color? idleColor;
+  final Color? normalColor;
+
+  ///dim Color
+  final Color? dimColor;
 
   @override
   State<CrossScroll> createState() => _CrossScrollState();
@@ -113,24 +98,41 @@ class _CrossScrollState extends State<CrossScroll> {
   // final StreamController<Thumb> _streamController = StreamController<Thumb>();
 
   double? width = 0;
-
-  double? fullHeight = 0;
-
-  final _trackKey = GlobalKey();
-
-  bool _onHoverTrack = false;
-
-  double? thumbCurrentPosition = 0;
+  double? height = 0;
   double previousWidth = 0;
-  double _viewport = 0;
-  double _maxExtent = 0.1;
-  double _thumbWidth = 55;
+  double previousHeight = 0;
+  double? horizontalFullSize = 0;
+  double? verticalFullSize = 0;
+
+  final _horizontalTrackKey = GlobalKey();
+  final _verticalTrackKey = GlobalKey();
+
+  bool _onHoverHorizontalTrack = false;
+  bool _onHoverVerticalTrack = false;
+
+  double? horizontalThumbCurrentPosition = 0;
+  double? verticalThumbCurrentPosition = 0;
+
+  double _horizontalViewPort = 0;
+  double _verticalViewPort = 0;
+  double _horizontalMaxExtent = 0.1;
+  double _verticalMaxExtent = 0.1;
+  double _horizontalThumbWidth = 55;
+  double _verticalThumbWidth = 55;
 
   ///This method update thumb position while scrolling,
   /// Platforms [Android and IOS]
-  updateThumbPositionWithScroll() {
+  updateThumbPositionWithScroll(Axis orientation) {
+    ScrollController controller = getController(orientation);
+    // print(controller.position.pixels);
     setState(() {
-      thumbCurrentPosition = horizontalScrollController.offset / ratio;
+      if (orientation == Axis.horizontal) {
+        horizontalThumbCurrentPosition =
+            (controller.offset) / ratio(orientation);
+      } else {
+        verticalThumbCurrentPosition =
+            (controller.position.pixels) / ratio(orientation);
+      }
     });
   }
 
@@ -140,10 +142,17 @@ class _CrossScrollState extends State<CrossScroll> {
     super.initState();
     widget.horizontalScrollController != null
         ? widget.horizontalScrollController
-            ?.addListener(() => updateThumbPositionWithScroll())
+            ?.addListener(() => updateThumbPositionWithScroll(Axis.horizontal))
         : horizontalScrollController
-            .addListener(() => updateThumbPositionWithScroll());
+            .addListener(() => updateThumbPositionWithScroll(Axis.horizontal));
+
+    widget.verticalScrollController != null
+        ? widget.verticalScrollController
+            ?.addListener(() => updateThumbPositionWithScroll(Axis.vertical))
+        : verticalScrollController
+            .addListener(() => updateThumbPositionWithScroll(Axis.vertical));
   }
+
 
   @override
   void dispose() {
@@ -152,24 +161,31 @@ class _CrossScrollState extends State<CrossScroll> {
     widget.horizontalScrollController != null
         ? widget.horizontalScrollController?.dispose()
         : horizontalScrollController.dispose();
+    widget.verticalScrollController != null
+        ? widget.verticalScrollController?.dispose()
+        : verticalScrollController.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     width = MediaQuery.of(context).size.width;
+    height = MediaQuery.of(context).size.height;
     // setState(() {
     //   print(horizontalScrollController.offset-_thumbSize);
     //
     // });
 
 // thumbCurrentPosition=widget.horizontalScrollController?.offset??horizontalScrollController.offset;
-    if (width != previousWidth) {
+    if (width != previousWidth || height != previousHeight) {
       Future.delayed(Duration.zero, () {
         setState(() {
-          initializeControllerValues();
-          keepThumbInRangeWhileResizingScreen();
+          initializeControllerValues(Axis.vertical);
+          initializeControllerValues(Axis.horizontal);
+          keepThumbInRangeWhileResizingScreen(Axis.vertical);
+          keepThumbInRangeWhileResizingScreen(Axis.horizontal);
         });
       });
+      previousHeight = height!;
       previousWidth = width!;
     }
 
@@ -179,25 +195,14 @@ class _CrossScrollState extends State<CrossScroll> {
     CrossScrollDesign? hStyle = widget.horizontalScroll;
     return Scaffold(
       body: Stack(
+        alignment: Alignment.bottomRight,
         children: [
-          CustomCrossScrollBar(
-            controller:
-                widget.verticalScrollController ?? verticalScrollController,
-            isAlwaysShown: vBar?.isAlwaysShown,
-            // interactive: vBar.interactive,
-            trackVisibility: vBar?.trackVisibility,
-            showTrackOnHover: vBar?.showTrackOnHover,
-            hoverThickness: vBar?.hoverThickness,
-            thickness: vBar?.thickness,
-            idleColor: widget.idleColor,
-            hoverColor: widget.hoverColor,
-            radius: vBar?.radius,
-            notificationPredicate: vBar?.notificationPredicate,
+          ScrollConfiguration(
+            behavior:
+                ScrollConfiguration.of(context).copyWith(scrollbars: false),
             child: SingleChildScrollView(
-              reverse: vStyle?.reverse ?? false,
               clipBehavior: vStyle?.clipBehavior ?? Clip.hardEdge,
-              dragStartBehavior:
-                  vStyle?.dragStartBehavior ?? DragStartBehavior.start,
+              dragStartBehavior:DragStartBehavior.start,
               keyboardDismissBehavior: vStyle?.keyboardDismissBehavior ??
                   ScrollViewKeyboardDismissBehavior.manual,
               restorationId: vStyle?.restorationId,
@@ -208,10 +213,8 @@ class _CrossScrollState extends State<CrossScroll> {
               controller:
                   widget.verticalScrollController ?? verticalScrollController,
               child: SingleChildScrollView(
-                reverse: hStyle?.reverse ?? false,
                 clipBehavior: hStyle?.clipBehavior ?? Clip.hardEdge,
-                dragStartBehavior:
-                    hStyle?.dragStartBehavior ?? DragStartBehavior.start,
+                dragStartBehavior:DragStartBehavior.start,
                 keyboardDismissBehavior: hStyle?.keyboardDismissBehavior ??
                     ScrollViewKeyboardDismissBehavior.manual,
                 restorationId: hStyle?.restorationId,
@@ -226,104 +229,485 @@ class _CrossScrollState extends State<CrossScroll> {
             ),
           ),
 
-          ///Custom Scroll track
-          _viewport == _maxExtent
+          ///Horizontal Scroll track
+          _horizontalViewPort == _horizontalMaxExtent
               ? const SizedBox(
                   width: 0,
                   height: 0,
                 )
-              : _track(bar: hBar!),
+              : _scrollTrack(
+                  bar: hBar!,
+                  orientation: Axis.horizontal,
+                ),
+          // _track(bar: hBar!, orientation: Axis.horizontal),
 
-          ///Custom scroll thumb
-          _viewport == _maxExtent
+          ///Horizontal scroll thumb
+          _horizontalViewPort == _horizontalMaxExtent
               ? const SizedBox(
                   width: 0,
                   height: 0,
                 )
-              : _thumbButton(thumb: hBar!),
+              : _scrollThumb(thumb: hBar!, orientation: Axis.horizontal),
+
+          ///Vertical Scroll track
+          _verticalViewPort == _verticalMaxExtent
+              ? const SizedBox(
+                  width: 0,
+                  height: 0,
+                )
+              : _scrollTrack(
+                  bar: vBar!,
+                  orientation: Axis.vertical,
+                ),
+
+          ///Vertical scroll thumb
+          _verticalViewPort == _verticalMaxExtent
+              ? const SizedBox(
+                  width: 0,
+                  height: 0,
+                )
+              : _scrollThumb(thumb: vBar!, orientation: Axis.vertical),
         ],
       ),
     );
   }
 
   /// get  ratio by fullSize/(visibleSize-thumbSize)
-  double get ratio => (horizontalScrollController.position.maxScrollExtent /
-      (horizontalScrollController.position.viewportDimension - _thumbSize));
+  double ratio(Axis orientation) {
+    ScrollController? controller;
+    if (orientation == Axis.horizontal) {
+      controller =
+          widget.horizontalScrollController ?? horizontalScrollController;
+    } else {
+      controller = widget.verticalScrollController ?? verticalScrollController;
+    }
+    return (controller.position.maxScrollExtent /
+        (controller.position.viewportDimension - _thumbSize(orientation)));
+  }
 
   ///get ThumbSize = (viewPortSize / maxExtent - minExtent + viewPortSize) * trackLength
-  double get _thumbSize {
-    double size = (horizontalScrollController.position.viewportDimension /
-            (horizontalScrollController.position.maxScrollExtent -
-                horizontalScrollController.position.minScrollExtent +
-                horizontalScrollController.position.viewportDimension)) *
-        horizontalScrollController.position.viewportDimension;
-    if (size < 55) {
+  double _thumbSize(Axis orientation) {
+    ScrollController? _controller = getController(orientation);
+
+    double _size = (_controller.position.viewportDimension /
+            (_controller.position.maxScrollExtent -
+                _controller.position.minScrollExtent +
+                _controller.position.viewportDimension)) *
+        _controller.position.viewportDimension;
+
+    if (_size < 55) {
       return 55.0;
     } else {
-      return size;
+      return _size;
     }
   }
 
-  double get viewPort => horizontalScrollController.position.viewportDimension;
+  double viewPort(Axis orientation) {
+    ScrollController? _controller = getController(orientation);
+    return _controller.position.viewportDimension;
+  }
 
-  double get virtualTotalSize =>
-      horizontalScrollController.position.maxScrollExtent;
+  double virtualTotalSize(Axis orientation) {
+    ScrollController? _controller = getController(orientation);
+    return _controller.position.maxScrollExtent;
+  }
 
-  double jumpTo(delta) => (delta * ratio) + horizontalScrollController.offset;
+  double jumpTo(delta, Axis orientation) {
+    ScrollController _controller = getController(orientation);
+    return (delta * ratio) + _controller.offset;
+  }
 
-  double get trackLength => _trackKey.currentContext!.size!.width;
+  double trackLength(Axis orientation) {
+    return orientation == Axis.horizontal
+        ? _horizontalTrackKey.currentContext!.size!.width
+        : _verticalTrackKey.currentContext!.size!.width;
+  }
 
-  @Deprecated(' Will stop  working from next version')
-  bool get isControllerLoaded =>
-      horizontalScrollController.position.maxScrollExtent.toString().isNotEmpty
-          ? true
-          : false;
 
-  void _onDrag(DragUpdateDetails a) {
+  void _onDrag(DragUpdateDetails a, {required Axis orientation}) {
     // if (kDebugMode) {
     //   // print('thumbCurrentPosition: ${thumbCurrentPosition!}');
     // }
-    setState(() {
-      if (0 <= thumbCurrentPosition! &&
-          thumbCurrentPosition! <= viewPort - _thumbSize) {
-        thumbCurrentPosition = thumbCurrentPosition! + a.delta.dx;
-        if (thumbCurrentPosition! != 0 ||
-            thumbCurrentPosition! != viewPort - _thumbSize) {
-          horizontalScrollController.jumpTo(thumbCurrentPosition! * ratio);
+    if (orientation == Axis.horizontal) {
+      ///horizontal
+
+      if (0 > horizontalThumbCurrentPosition!) {
+        horizontalThumbCurrentPosition = 0;
+      } else if (horizontalThumbCurrentPosition! >
+          viewPort(orientation) - _thumbSize(orientation)) {
+        horizontalThumbCurrentPosition =
+            viewPort(orientation) - _thumbSize(orientation);
+      } else if (0 <= horizontalThumbCurrentPosition! &&
+          horizontalThumbCurrentPosition! <=
+              viewPort(orientation) - _thumbSize(orientation)) {
+        horizontalThumbCurrentPosition =
+            horizontalThumbCurrentPosition! + a.delta.dx;
+        if (horizontalThumbCurrentPosition! != 0 ||
+            horizontalThumbCurrentPosition! !=
+                viewPort(orientation) - _thumbSize(orientation)) {
+          getController(orientation)
+              .jumpTo(horizontalThumbCurrentPosition! * ratio(orientation));
           // if (kDebugMode) {
           //   print('JumpTo:${thumbCurrentPosition! * ratio}');
           // }
         }
       }
-      if (0 > thumbCurrentPosition!) {
-        thumbCurrentPosition = 0;
-      } else if (thumbCurrentPosition! > viewPort - _thumbSize) {
-        thumbCurrentPosition = viewPort - _thumbSize;
+    } else {
+      ///vertical
+      print(a.delta.dy);
+      if (verticalThumbCurrentPosition! < 0) {
+        verticalThumbCurrentPosition = 0;
+      } else if (verticalThumbCurrentPosition! >
+          viewPort(orientation) - _thumbSize(orientation)) {
+        verticalThumbCurrentPosition =
+            viewPort(orientation) - _thumbSize(orientation);
+      } else if (verticalThumbCurrentPosition! >= 0 &&
+          verticalThumbCurrentPosition! <=
+              viewPort(orientation) - _thumbSize(orientation)) {
+        verticalThumbCurrentPosition =
+            verticalThumbCurrentPosition! + a.delta.dy;
+        if (verticalThumbCurrentPosition! != 0 ||
+            verticalThumbCurrentPosition! !=
+                viewPort(orientation) - _thumbSize(orientation)) {
+          getController(orientation)
+              .jumpTo(verticalThumbCurrentPosition! * ratio(orientation));
+        }
       }
-    });
+    }
+  }
+  void onTrackClick(TapDownDetails tapDownDetails,
+      {required Axis orientation}) {
+    print(tapDownDetails.localPosition.dy);
+    // print(verticalThumbCurrentPosition);
+    // print(_verticalThumbWidth);
+
+    ScrollController _hController =
+        widget.horizontalScrollController ?? horizontalScrollController;
+    ScrollController _vController =
+        widget.verticalScrollController ?? verticalScrollController;
+
+    double _horizontalTrackLengthFromThumbPoint =
+        horizontalThumbCurrentPosition! + _horizontalThumbWidth;
+    double _verticalTrackLengthFromThumbPoint =
+        verticalThumbCurrentPosition! + _verticalThumbWidth;
+
+    if (orientation == Axis.horizontal) {
+      ///horizontal track
+      if (tapDownDetails.localPosition.dx > horizontalThumbCurrentPosition!) {
+        if (_horizontalTrackLengthFromThumbPoint <
+            _horizontalViewPort - _horizontalThumbWidth) {
+          // print('+ive\nless than _viewport - _thumbWidth');
+
+          horizontalThumbCurrentPosition =
+              horizontalThumbCurrentPosition! + _horizontalThumbWidth;
+          _hController.animateTo(
+              horizontalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        } else {
+          horizontalThumbCurrentPosition =
+              _horizontalViewPort - _horizontalThumbWidth;
+          _hController.animateTo(
+              horizontalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        }
+      } else if (tapDownDetails.localPosition.dx <
+          horizontalThumbCurrentPosition!) {
+        if (horizontalThumbCurrentPosition! > _horizontalThumbWidth) {
+          // print('_crnt$_crnt');
+          // print('_thumbWidth$_thumbWidth');
+          // print('-ive\ngreater than thumb');
+          horizontalThumbCurrentPosition =
+              horizontalThumbCurrentPosition! - _horizontalThumbWidth;
+          _hController.animateTo(
+              horizontalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        } else if (horizontalThumbCurrentPosition! < _horizontalThumbWidth) {
+          // print('-ive\nless than thumb');
+          horizontalThumbCurrentPosition = 0;
+          _hController.animateTo(
+              horizontalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        }
+      }
+    } else {
+      ///vertical track
+      if (tapDownDetails.localPosition.dy > verticalThumbCurrentPosition!) {
+        if (_verticalTrackLengthFromThumbPoint <
+            _verticalViewPort - _verticalThumbWidth) {
+          // print('+ive\nless than _viewport - _thumbWidth');
+
+          verticalThumbCurrentPosition =
+              verticalThumbCurrentPosition! + _verticalThumbWidth;
+          _vController.animateTo(
+              verticalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        } else {
+          verticalThumbCurrentPosition =
+              _verticalViewPort - _verticalThumbWidth;
+          _vController.animateTo(
+              verticalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        }
+      } else if (tapDownDetails.localPosition.dy <
+          verticalThumbCurrentPosition!) {
+        if (verticalThumbCurrentPosition! > _verticalThumbWidth) {
+          // print('_crnt$_crnt');
+          // print('_thumbWidth$_thumbWidth');
+          // print('-ive\ngreater than thumb');
+          verticalThumbCurrentPosition =
+              verticalThumbCurrentPosition! - _verticalThumbWidth;
+          _vController.animateTo(
+              verticalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        } else if (verticalThumbCurrentPosition! < _verticalThumbWidth) {
+          // print('-ive\nless than thumb');
+          verticalThumbCurrentPosition = 0;
+          _vController.animateTo(
+              verticalThumbCurrentPosition! * ratio(orientation),
+              duration: const Duration(
+                milliseconds: 300,
+              ),
+              curve: Curves.linear);
+        }
+      }
+    }
   }
 
-  Widget _thumbButton({required CrossScrollBar thumb}) {
-    double? _height() =>
-        thumb.trackVisibility == true || thumb.showTrackOnHover == true
-            ? _onHoverTrack
-                ? thumb.hoverThickness
-                : thumb.thickness
-            : thumb.thickness;
+  void keepThumbInRangeWhileResizingScreen(Axis orientation) {
+    bool _isHorizontal = orientation == Axis.horizontal ? true : false;
+    double _thumbCurrentPosition = _isHorizontal
+        ? horizontalThumbCurrentPosition!
+        : verticalThumbCurrentPosition!;
+    if (_thumbCurrentPosition >
+        viewPort(orientation) - _thumbSize(orientation)) {
+      _thumbCurrentPosition = viewPort(orientation) - _thumbSize(orientation);
+      getController(orientation)
+          .jumpTo(_thumbCurrentPosition * ratio(orientation));
+    }
+  }
 
-    Widget posi = Positioned(
-      bottom: 1,
-      left: thumbCurrentPosition,
+  void initializeControllerValues(Axis orientation) {
+    bool _isHorizontal = orientation == Axis.horizontal ? true : false;
+    ScrollController _controller = getController(orientation);
+    if (_isHorizontal) {
+      ///horizontal
+      horizontalFullSize = _controller.position.maxScrollExtent;
+      _horizontalMaxExtent = (_controller.position.viewportDimension +
+          _controller.position.maxScrollExtent);
+      _horizontalViewPort = _controller.position.viewportDimension;
+      _horizontalThumbWidth = _thumbSize(orientation);
+    } else {
+      ///Vertical
+      verticalFullSize = _controller.position.maxScrollExtent;
+      _verticalMaxExtent = (_controller.position.viewportDimension +
+          _controller.position.maxScrollExtent);
+      _verticalViewPort = _controller.position.viewportDimension;
+      _verticalThumbWidth = _thumbSize(orientation);
+    }
+  }
+
+  ScrollController getController(Axis orientation) {
+    return orientation == Axis.vertical
+        ? widget.verticalScrollController ?? verticalScrollController
+        : widget.horizontalScrollController ?? horizontalScrollController;
+  }
+
+  BorderRadiusGeometry _thumbRadius(
+      {required CrossScrollBar bar,
+      required bool hover,
+      required Axis orientation}) {
+    return BorderRadius.all(Radius.elliptical(
+        bar.thumbRadius?.x ?? bar.thickness ?? 8,
+        bar.thumbRadius?.y ?? bar.thickness ?? 8));
+  }
+
+  Color _thumbColor(
+    BuildContext context, {
+    required CrossScrollBar bar,
+    required bool hover,
+    required Axis orientation,
+  }) {
+    ///called by Cross scroll constructor
+    Color normalColor =
+        widget.normalColor ?? const Color.fromARGB(228, 26, 26, 26);
+    Color dimColor = widget.dimColor ?? const Color.fromARGB(159, 24, 24, 24);
+
+    if (bar.thumb == ScrollThumb.hoverShow ||
+        bar.track == ScrollTrack.onHover) {
+      return hover ? widget.hoverColor ?? normalColor : dimColor;
+    } else if (bar.thumb == ScrollThumb.alwaysDim) {
+      return dimColor;
+    } else if (bar.thumb == ScrollThumb.alwaysShow) {
+      return normalColor;
+    } else {
+      return normalColor;
+    }
+  }
+
+  ///if no color provided, default colors will be assign
+  ///
+  ///[trackColor]??Colors.black.withOpacity(0.03);
+  ///
+  ///[noColor]??Colors.transparent;
+  ///
+  ///[borderColor]??Colors.black.withOpacity(0.0851);
+
+  Widget _scrollTrack(
+      {required CrossScrollBar bar,
+      required Axis orientation,
+      Color? trackColor,
+      Color? noColor,
+      Color? borderColor}) {
+    bool? _isHorizontal = Axis.horizontal == orientation ? true : false;
+
+    Color _trackColor = trackColor ?? Colors.black.withOpacity(0.03);
+    Color _noColor = noColor ?? Colors.transparent;
+    Color _borderColor = borderColor ?? Colors.black.withOpacity(0.0851);
+
+    Widget _trackWidget({
+      bool isOnTapScroll = true,
+      required Color color,
+      required Color borderColor,
+      double padding = 2,
+    }) =>
+        GestureDetector(
+          onTapDown: (tapDown) => isOnTapScroll
+              ? onTrackClick(tapDown, orientation: orientation)
+              : null,
+          child: InkWell(
+            onHover: (hover) => setState(() => _isHorizontal
+                ? _onHoverHorizontalTrack = hover
+                : _onHoverVerticalTrack = hover),
+            onTap: () {},
+            child: Container(
+              alignment: Alignment.center,
+              padding: EdgeInsets.symmetric(
+                  vertical: _isHorizontal ? padding : 0,
+                  horizontal: !_isHorizontal ? padding : 0),
+              decoration: BoxDecoration(
+                  color: color,
+                  border: Border(
+                      left: !_isHorizontal
+                          ? BorderSide(width: 0.8, color: borderColor)
+                          : BorderSide(width: 0, color: Colors.transparent),
+                      top: _isHorizontal
+                          ? BorderSide(width: 0.8, color: borderColor)
+                          : BorderSide(width: 0, color: Colors.transparent))),
+              key: _isHorizontal ? _horizontalTrackKey : _verticalTrackKey,
+              width: !_isHorizontal
+                  ? _onHoverVerticalTrack
+                      ? bar.hoverThickness! + padding * 1.4
+                      : bar.thickness! + padding * 1.4
+                  : null,
+              height: _isHorizontal
+                  ? _onHoverHorizontalTrack
+                      ? bar.hoverThickness! + padding *  1.4
+                      : bar.thickness! + padding *  1.4
+                  : null,
+            ),
+          ),
+        );
+
+    /// orientation switch
+    switch (orientation) {
+      case Axis.vertical:
+
+        ///track bar switch
+        switch (bar.track) {
+          case ScrollTrack.onHover:
+            return _trackWidget(
+              color: _onHoverVerticalTrack ? _trackColor : _noColor,
+              borderColor: _onHoverVerticalTrack ? _borderColor : _noColor,
+            );
+
+          case ScrollTrack.show:
+            return _trackWidget(
+              color: _trackColor,
+              borderColor: _borderColor,
+            );
+          case ScrollTrack.hidden:
+            return _trackWidget(
+              color: _noColor,
+              borderColor: _noColor,
+            );
+        }
+
+      case Axis.horizontal:
+        switch (bar.track) {
+          case ScrollTrack.onHover:
+            return _trackWidget(
+              color: _onHoverHorizontalTrack ? _trackColor : _noColor,
+              borderColor: _onHoverHorizontalTrack ? _borderColor : _noColor,
+            );
+
+          case ScrollTrack.show:
+            return _trackWidget(
+              color: _trackColor,
+              borderColor: _borderColor,
+            );
+          case ScrollTrack.hidden:
+            return _trackWidget(
+              color: _noColor,
+              borderColor: _noColor,
+            );
+        }
+    }
+  }
+
+  Widget _scrollThumb(
+      {required CrossScrollBar thumb, required Axis orientation}) {
+    bool _isHorizontal = orientation == Axis.horizontal ? true : false;
+    bool _trackHover =
+        _isHorizontal ? _onHoverHorizontalTrack : _onHoverVerticalTrack;
+    double? _thumbThickness() =>
+        _trackHover ? thumb.hoverThickness : thumb.thickness;
+
+    Widget position = Positioned(
+      top: _isHorizontal ? null : verticalThumbCurrentPosition,
+      bottom: _isHorizontal ? 1 : null,
+      left: !_isHorizontal ? null : horizontalThumbCurrentPosition,
+      right: !_isHorizontal?1:null,
       child: SizedBox(
-          height: _height(),
-          width: _thumbWidth < 55 ? 55 : _thumbWidth,
+          height: _isHorizontal
+              ? _thumbThickness()
+              : _verticalThumbWidth < 55
+                  ? 55
+                  : _verticalThumbWidth,
+          width: !_isHorizontal
+              ? _thumbThickness()
+              : _horizontalThumbWidth < 55
+                  ? 55
+                  : _horizontalThumbWidth,
           child: GestureDetector(
             // onTapDown: (tip) =>fingerTip= tip.localPosition.dx,
-            onHorizontalDragUpdate: _onDrag,
+            onPanUpdate: (drag) => _onDrag(drag, orientation: orientation),
             child: InkWell(
               onHover: (hover) {
                 setState(() {
-                  _onHoverTrack = hover;
+                  _isHorizontal
+                      ? _onHoverHorizontalTrack = hover
+                      : _onHoverVerticalTrack = hover;
                 });
               },
               autofocus: true,
@@ -332,265 +716,24 @@ class _CrossScrollState extends State<CrossScroll> {
               onTap: () {},
               child: Container(
                 decoration: BoxDecoration(
-                  borderRadius: _thumbRadius(bar: thumb, hover: _onHoverTrack),
-                  color: _thumbColor(bar: thumb, hover: _onHoverTrack),
+                  borderRadius: _thumbRadius(
+                      bar: thumb,
+                      hover: _isHorizontal
+                          ? _onHoverHorizontalTrack
+                          : _onHoverVerticalTrack,
+                      orientation: orientation),
+                  color: _thumbColor(context,
+                      bar: thumb,
+                      hover: _isHorizontal
+                          ? _onHoverHorizontalTrack
+                          : _onHoverVerticalTrack,
+                      orientation: orientation),
                 ),
               ),
             ),
           )),
     );
 
-    return posi;
-  }
-
-  BorderRadiusGeometry _thumbRadius(
-      {required CrossScrollBar bar, required bool hover}) {
-    return BorderRadius.all(Radius.elliptical(
-        bar.radius?.x ?? (bar.thickness ?? 14),
-        bar.radius?.y ?? (bar.thickness ?? 14)));
-  }
-
-  Widget _track({
-    required CrossScrollBar bar,
-  }) {
-    bool? _isTV = bar.trackVisibility;
-    bool? _isSTOnHover = bar.showTrackOnHover;
-    bool? _isAlwaysShown = bar.isAlwaysShown;
-
-    Widget _value;
-    Widget _trackWidget({
-      bool isOnTapScroll = true,
-      required Color color,
-      required Color borderColor,
-      double paddingVert = 2,
-    }) =>
-        Align(
-          alignment: Alignment.bottomCenter,
-          child: GestureDetector(
-            onTapDown: isOnTapScroll ? onTrackClick : null,
-            child: InkWell(
-              onHover: (hover) => setState(() => _onHoverTrack = hover),
-              onTap: () {},
-              child: Container(
-                alignment: Alignment.center,
-                // margin: const EdgeInsets.symmetric(vertical: 3.0),
-                padding: EdgeInsets.symmetric(vertical: paddingVert),
-                decoration: BoxDecoration(
-                    color: color,
-                    border: Border(
-                        top: BorderSide(width: 0.8, color: borderColor))),
-                key: _trackKey,
-
-                height: _onHoverTrack
-                    ? bar.hoverThickness! + 3
-                    : bar.thickness! + 3,
-              ),
-            ),
-          ),
-        );
-    if (_isTV!) {
-      ///Always Visible
-      _value = _trackWidget(
-        borderColor: Colors.black.withOpacity(0.0851),
-        color: bar.trackVisibility != null
-            ? bar.trackVisibility!
-                ? Colors.black.withOpacity(0.03)
-                : Colors.black.withOpacity(0.00)
-            : Colors.transparent,
-      );
-    } else {
-      if (_isSTOnHover!) {
-        ///Visible On Hover
-        ///
-        _value = _trackWidget(
-            color: _onHoverTrack
-                ? Colors.black.withOpacity(0.03)
-                : Colors.transparent,
-            borderColor: _isAlwaysShown!
-                ? _onHoverTrack
-                    ? Colors.black.withOpacity(0.0851)
-                    : Colors.black.withOpacity(0.00)
-                : _onHoverTrack
-                    ? Colors.black.withOpacity(0.0851)
-                    : Colors.black.withOpacity(0.00));
-      } else {
-        ///Don't Show Track
-        _value = _trackWidget(
-          color: const Color.fromARGB(0, 0, 0, 0),
-          borderColor: const Color.fromARGB(0, 0, 0, 0),
-        );
-      }
-    }
-    return _value;
-  }
-
-  Color _thumbColor({
-    required CrossScrollBar bar,
-    required bool hover,
-    Color fullColor = const Color.fromARGB(239, 16, 16, 16),
-    // Color middleColor = const Color.fromARGB(155, 17, 17, 17),
-  }) {
-    bool trackVisibility = bar.trackVisibility ?? false;
-
-    Color? color;
-    if (bar.isAlwaysShown != null) {
-      bar.isAlwaysShown!
-          ? trackVisibility
-              ?
-
-              ///When [trackVisibility] is true
-              ///The hover will be ignored
-
-              color = widget.hoverColor ?? fullColor
-              :
-
-              ///When [trackVisibility] is false
-              hover
-
-                  ///Show Thumb Full Color on Hover when [isAlwaysShown]  is true
-                  ? color = widget.hoverColor ?? fullColor
-
-                  ///Show Half Color When not Hovered when [isAlwaysShown]  is true
-                  : kIsWeb
-                      ? color = widget.idleColor ??
-                          const Color.fromARGB(118, 17, 17, 17)
-                      : Platform.isAndroid
-                          ? color = widget.idleColor ??
-                              Theme.of(context).highlightColor.withOpacity(0.9)
-                          : color = widget.idleColor ??
-                              const Color.fromARGB(118, 17, 17, 17)
-          : hover
-
-              ///Show Thumb Full Color on Hover when [isAlwaysShown]  is false
-              ? color = widget.hoverColor ?? fullColor
-
-              ///Don't show Thumb when not Hovered
-              : color = const Color(0x00111111);
-    } else {
-      hover
-
-          ///Show Thumb Full Color on Hover
-          ? color = fullColor
-
-          ///Don't show Thumb when not Hovered
-          : color = const Color(0x00111111);
-    }
-    return color;
-  }
-
-  // Color _thumbColorNew({
-  //   required CrossScrollBar bar,
-  //   required bool hover,
-  //   Color fullColor = const Color.fromARGB(239, 16, 16, 16),
-  //   // Color middleColor = const Color.fromARGB(155, 17, 17, 17),
-  // }) {
-  //   bool trackVisibility = bar.trackVisibility ?? false;
-  //
-  //   Color? color;
-  //   if (bar.isAlwaysShown != null) {
-  //     bar.isAlwaysShown!
-  //         ? trackVisibility
-  //             ?
-  //
-  //             ///When [trackVisibility] is true
-  //             ///The hover will be ignored
-  //
-  //             color = widget.hoverColor ?? fullColor
-  //             :
-  //
-  //             ///When [trackVisibility] is false
-  //             hover
-  //
-  //                 ///Show Thumb Full Color on Hover when [isAlwaysShown]  is true
-  //                 ? color = widget.hoverColor ?? fullColor
-  //
-  //                 ///Show Half Color When not Hovered when [isAlwaysShown]  is true
-  //                 : Platform.isAndroid || Platform.isIOS
-  //                     ? color = widget.idleColor ??
-  //                         Theme.of(context).highlightColor.withOpacity(0.9)
-  //                     : color = widget.idleColor ??
-  //                         const Color.fromARGB(118, 17, 17, 17)
-  //         : hover
-  //
-  //             ///Show Thumb Full Color on Hover when [isAlwaysShown]  is false
-  //             ? color = widget.hoverColor ?? fullColor
-  //
-  //             ///Don't show Thumb when not Hovered
-  //             : color = const Color(0x00111111);
-  //   } else {
-  //     hover
-  //
-  //         ///Show Thumb Full Color on Hover
-  //         ? color = fullColor
-  //
-  //         ///Don't show Thumb when not Hovered
-  //         : color = const Color(0x00111111);
-  //   }
-  //   return color;
-  // }
-
-  void onTrackClick(TapDownDetails tapDownDetails) {
-    if (kDebugMode) {
-      // print('thumbCurrentPosition: $thumbCurrentPosition');
-      // print('tapDownDetails.dx: ${tapDownDetails.localPosition.dx}');
-    }
-    double _trackLengthfromThumbPoint = thumbCurrentPosition! + _thumbWidth;
-
-    setState(() {
-      if (tapDownDetails.localPosition.dx > thumbCurrentPosition!) {
-        if (_trackLengthfromThumbPoint < _viewport - _thumbWidth) {
-          // print('+ive\nless than _viewport - _thumbWidth');
-
-          thumbCurrentPosition = thumbCurrentPosition! + _thumbWidth;
-          horizontalScrollController.animateTo(thumbCurrentPosition! * ratio,
-              duration: const Duration(
-                milliseconds: 300,
-              ),
-              curve: Curves.linear);
-        } else {
-          thumbCurrentPosition = _viewport - _thumbWidth;
-          horizontalScrollController.animateTo(thumbCurrentPosition! * ratio,
-              duration: const Duration(
-                milliseconds: 300,
-              ),
-              curve: Curves.linear);
-        }
-      } else if (tapDownDetails.localPosition.dx < thumbCurrentPosition!) {
-        if (thumbCurrentPosition! > _thumbWidth) {
-          // print('_crnt$_crnt');
-          // print('_thumbWidth$_thumbWidth');
-          // print('-ive\ngreater than thumb');
-          thumbCurrentPosition = thumbCurrentPosition! - _thumbWidth;
-          horizontalScrollController.animateTo(thumbCurrentPosition! * ratio,
-              duration: const Duration(
-                milliseconds: 300,
-              ),
-              curve: Curves.linear);
-        } else if (thumbCurrentPosition! < _thumbWidth) {
-          // print('-ive\nless than thumb');
-          thumbCurrentPosition = 0;
-          horizontalScrollController.animateTo(thumbCurrentPosition! * ratio,
-              duration: const Duration(
-                milliseconds: 300,
-              ),
-              curve: Curves.linear);
-        }
-      }
-    });
-  }
-
-  void keepThumbInRangeWhileResizingScreen() {
-    if (thumbCurrentPosition! > viewPort - _thumbSize) {
-      thumbCurrentPosition = viewPort - _thumbSize;
-      horizontalScrollController.jumpTo(thumbCurrentPosition! * ratio);
-    }
-  }
-
-  void initializeControllerValues() {
-    fullHeight = horizontalScrollController.position.maxScrollExtent;
-    _maxExtent = (horizontalScrollController.position.viewportDimension +
-        horizontalScrollController.position.maxScrollExtent);
-    _viewport = horizontalScrollController.position.viewportDimension;
-    _thumbWidth = _thumbSize;
+    return position;
   }
 }