+++ lib/src/cross_scroll_decor.dart
@@ -1,40 +1,15 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 class CrossScrollDesign {
   CrossScrollDesign({
-    this.reverse = false,
     this.padding,
-    // this.primary,
     this.physics,
-    this.dragStartBehavior = DragStartBehavior.start,
     this.clipBehavior = Clip.hardEdge,
     this.restorationId,
     this.keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
   });
-  //
-  // final bool reverse;
-  // final EdgeInsetsGeometry? padding;
-  // final bool? primary;
-  // final ScrollPhysics? physics;
-  // final DragStartBehavior? dragStartBehavior;
-  // final Clip? clipBehavior;
-  // final String? restorationId;
-  // final ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior;
 
-  /// Whether the scroll view scrolls in the reading direction.
-  ///
-  /// For example, if the reading direction is left-to-right and
-  /// [scrollDirection] is [Axis.horizontal], then the scroll view scrolls from
-  /// left to right when [reverse] is false and from right to left when
-  /// [reverse] is true.
-  ///
-  /// Similarly, if [scrollDirection] is [Axis.vertical], then the scroll view
-  /// scrolls from top to bottom when [reverse] is false and from bottom to top
-  /// when [reverse] is true.
-  ///
-  /// Defaults to false.
-  final bool reverse;
+
 
   /// The amount of space by which to inset the child.
   final EdgeInsetsGeometry? padding;
@@ -46,9 +21,6 @@ class CrossScrollDesign {
   ///
   /// Defaults to matching platform conventions.
   final ScrollPhysics? physics;
-
-  /// {@macro flutter.widgets.scrollable.dragStartBehavior}
-  final DragStartBehavior dragStartBehavior;
 
   /// {@macro flutter.material.Material.clipBehavior}
   ///